Phase 2: migrate persistence layer to JPA (#17)
* refactor: migrate persistence layer to jpa

* docs: clarify jpa migration review notes

* docs: translate migration comments to chinese

+++ src/main/java/com/admin/base/component/ResponseInit.java
@@ -10,7 +10,7 @@ import com.admin.base.dto.response.system.LoginResponse;
  */
 public class ResponseInit {
 
-    public static LoginResponse initLoginResponse(Integer adminId,String userName,  String nickname, String token) {
+    public static LoginResponse initLoginResponse(Long adminId,String userName,  String nickname, String token) {
         LoginResponse loginResponse = new LoginResponse();
         loginResponse.setAccount(userName);
         loginResponse.setAdminId(adminId);
