+++ src/main/java/com/admin/base/config/security/TokenUser.java
@@ -12,7 +12,7 @@ public record TokenUser(
         String username,
         String password,
         String nickName,
-        Integer adminId,
+        Long adminId,
         List<String> perms
 ) {
     /**
